+++ apps/client/src/app/core/data/localized-data.service.ts
@@ -9,11 +9,13 @@ import { koActions } from './sources/ko-actions';
 import { mapIds } from './sources/map-ids';
 import { LazyDataService } from './lazy-data.service';
 import { Fate } from '../../pages/db/model/fate/fate';
+import { Npc } from '../../pages/db/model/npc/npc';
 import { Quest } from '../../pages/db/model/quest/quest';
 import { tripleTriadRules } from './sources/triple-triad-rules';
 import { zhActions } from './sources/zh-actions';
 import { ExtendedLanguageKeys } from './extended-language-keys';
 import { LazyData } from './lazy-data';
+import { Trait } from '../../pages/db/model/trait/trait';
 
 @Injectable()
 export class LocalizedDataService {
@@ -34,6 +36,10 @@ export class LocalizedDataService {
 
   public getInstanceName(id: number): any {
     return this.getRowWithExtendedLanguage('instances', id);
+  }
+
+  public getAchievementName(id: number): I18nName {
+    return this.getRowWithExtendedLanguage('achievements', id);
   }
 
   public getMapName(id: number): any {
@@ -60,8 +66,8 @@ export class LocalizedDataService {
     return row;
   }
 
-  public getNpc(id: number): I18nName {
-    return this.getRowWithExtendedLanguage('npcs', id);
+  public getNpc(id: number): Npc {
+    return this.getRowWithExtendedLanguage<Npc>('npcs', id);
   }
 
   public getLeve(id: number): I18nName {
@@ -106,8 +112,13 @@ export class LocalizedDataService {
     return row;
   }
 
-  public getTrait(id: number): any {
-    return this.getRowWithExtendedLanguage('traits', id);
+  public getTrait(id: number): Trait {
+    const row = this.getRowWithExtendedLanguage<Trait>('traits', id);
+    if (row && row.description) {
+      this.tryFillExtendedLanguage(row.description, id, { zhKey: 'zhTraitDescriptions', koKey: 'koTraitDescriptions' })
+    }
+
+    return row;
   }
 
   public getRaceName(id: number): any {
@@ -122,7 +133,7 @@ export class LocalizedDataService {
       ja: lazyRow.Name_ja,
       fr: lazyRow.Name_fr
     };
-    this.tryFillExtendedLanguage(row, id, { zhKey: 'tribes', koKey: 'tribes' });
+    this.tryFillExtendedLanguage(row, id, { zhKey: 'zhTribes', koKey: 'tribes' });
     return row;
   }
 
@@ -141,7 +152,7 @@ export class LocalizedDataService {
       ja: lazyRow.Name_ja,
       fr: lazyRow.Name_fr
     };
-    this.tryFillExtendedLanguage(row, id, { zhKey: 'baseParams', koKey: 'baseParams' });
+    this.tryFillExtendedLanguage(row, id, { zhKey: 'zhBaseParams', koKey: 'baseParams' });
 
     return row;
   }
@@ -256,8 +267,8 @@ export class LocalizedDataService {
     return array[id];
   }
 
-  private getRowWithExtendedLanguage(key: keyof LazyData, id: number | string): I18nName {
-    const row = this.getRow<I18nName>(this.lazyData.data[key], id);
+  private getRowWithExtendedLanguage<T extends I18nName = I18nName>(key: keyof LazyData, id: number | string): T {
+    const row = this.getRow<T>(this.lazyData.data[key], id);
     if (row === undefined) {
       return undefined;
     }
@@ -295,6 +306,20 @@ export class LocalizedDataService {
       const koRow = this.getRow(this.lazyData.data[koKey], id);
       row.ko = koRow !== undefined ? koRow.ko : row.en;
     }
+  }
+
+  public xivapiToI18n(value: any, key: any, fieldName = 'Name'): I18nName {
+    const row = {
+      en: value[`${fieldName}_en`],
+      fr: value[`${fieldName}_fr`],
+      de: value[`${fieldName}_de`],
+      ja: value[`${fieldName}_ja`],
+    };
+    
+    if (key !== null) {
+      this.tryFillExtendedLanguage(row, value.ID, this.guessExtendedLanguageKeys(key));
+    }
+    return row;
   }
 
   /**