+++ apps/client/src/app/pages/db/model/trait/trait.ts
@@ -1,0 +1,6 @@
+import { I18nName } from '../../../../model/common/i18n-name';
+
+export interface Trait extends I18nName {
+  icon: string;
+  description: I18nName;
+}