+++ apps/client/src/app/pages/simulator/rotation-tips/rotation-tips.module.ts
@@ -7,14 +7,11 @@ import { UseObserveBeforeFocused } from './tips/use-observe-before-focused';
 import { UseDurabilityRestorationLater } from './tips/use-durability-restoration-later';
 import { UsePatientTouchFaster } from './tips/use-patient-touch-faster';
 import { DoNotOverlapBuffs } from './tips/do-not-overlap-buffs';
-import { UseIngenuityBeforeByregot } from './tips/use-ingenuity-before-byregot';
 import { UseRapidSynthesisEarlier } from './tips/use-rapid-synthesis-earlier';
 import { UseMoreQualityActions } from './tips/use-more-quality-actions';
 import { UsePrudentTouchManipulation } from './tips/use-prudent-touch-manipulation';
 import { AvoidUsingGoodActions } from './tips/avoid-using-good-actions';
 import { DoNotUseBrandWithoutName } from './tips/do-not-use-brand-without-name';
-import { RequiresReuseToProc } from './tips/requires-reuse-to-proc';
-import { UseReuse } from './tips/use-reuse';
 
 export const ROTATION_TIPS = new InjectionToken('ROTATION_TIPS');
 
@@ -25,14 +22,11 @@ const tips: Class<RotationTip>[] = [
   UseDurabilityRestorationLater,
   UsePatientTouchFaster,
   DoNotOverlapBuffs,
-  UseIngenuityBeforeByregot,
   UseRapidSynthesisEarlier,
   UseMoreQualityActions,
   UsePrudentTouchManipulation,
   DoNotUseBrandWithoutName,
   AvoidUsingGoodActions,
-  RequiresReuseToProc,
-  UseReuse
 ];
 
 @NgModule({